+++ routes/actions.php
@@ -90,6 +90,7 @@ use CraftCms\Cms\Http\Controllers\Utilities\UtilitiesController;
 use CraftCms\Cms\Http\Middleware\EnsureTwoFactorChallengeIsRecent;
 use CraftCms\Cms\Http\Middleware\RequireAdmin;
 use CraftCms\Cms\Http\Middleware\RequireAdminChanges;
+use CraftCms\Cms\Http\Middleware\RequireConfirmedPassword;
 use CraftCms\Cms\Http\Middleware\RequireEdition;
 use CraftCms\Cms\Http\Middleware\RequireToken;
 use CraftCms\Cms\Http\Middleware\StartSessionWithoutPersistence;
@@ -376,12 +377,12 @@ Route::prefix($routes->cpActionTriggerRoutePrefix())->middleware(['craft.cp'])->
         Route::post('app/check-for-updates', [UpdatesController::class, 'check']);
         Route::post('app/cache-updates', [UpdatesController::class, 'cache']);
 
-        Route::middleware('password.confirm')->group(function () {
+        Route::middleware(RequireConfirmedPassword::class)->group(function () {
             Route::post('users/save-password', [PasswordController::class, 'store']);
         });
 
         Route::middleware([RequireEdition::class.':'.Edition::Team->value, 'can:editUsers'])->group(function () {
-            Route::middleware('password.confirm')->group(function () {
+            Route::middleware(RequireConfirmedPassword::class)->group(function () {
                 Route::post('users/impersonate', [ImpersonationController::class, 'impersonate']);
                 Route::post('users/get-impersonation-url', [ImpersonationController::class, 'getUrl']);
             });