+++ src/Site/Sites.php
@@ -35,7 +35,6 @@ use Illuminate\Container\Attributes\Scoped;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Date;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
@@ -43,6 +42,7 @@ use InvalidArgumentException;
 use Throwable;
 use Tpetry\QueryExpressions\Language\Alias;
 
+use function CraftCms\Cms\craftAuth;
 use function CraftCms\Cms\currentUser;
 use function CraftCms\Cms\maxPowerCaptain;
 
@@ -881,7 +881,7 @@ class Sites
     {
         $withDisabled ??= (
             app()->runningInConsole() ||
-            (request()->isCpRequest() && Auth::check())
+            (request()->isCpRequest() && craftAuth()->check())
         );
 
         return $withDisabled ? $this->allSitesById : $this->enabledSitesById;